refactor: 확장성을 위한 패키지 구조 변경
